Rename jog_arm->moveit_jog_arm
