+++ moveit_experimental/moveit_jog_arm/README.md
@@ -32,7 +32,7 @@ strictness: 2"
 
 Launch the jog node. This example uses commands from a SpaceNavigator joystick-like device:
 
-    roslaunch jog_arm spacenav_cpp.launch
+    roslaunch moveit_jog_arm spacenav_cpp.launch
 
 If you dont have a SpaceNavigator, send commands like this:
