+++ moveit_experimental/moveit_jog_arm/launch/spacenav_teleop_tools.launch
@@ -9,14 +9,14 @@
   <node name="spacenav_node" pkg="spacenav_node" type="spacenav_node" />
 
   <!-- This node does the jogging calculations -->
-  <node name="jog_arm_server" pkg="jog_arm" type="jog_arm_server" output="screen" >
+  <node name="jog_arm_server" pkg="moveit_jog_arm" type="jog_arm_server" output="screen" >
     <param name="parameter_ns" type="string" value="jog_arm_server" />
-    <rosparam command="load" file="$(find jog_arm)/config/ur_simulated_config.yaml" />
+    <rosparam command="load" file="$(find moveit_jog_arm)/config/ur_simulated_config.yaml" />
   </node>
 
   <!-- teleop_tools translates a joystick command into a twist message -->
   <arg name="cmd_vel" default="spacenav/joy" />
-  <arg name="teleop_config" default="$(find jog_arm)/config/spacenav_via_teleop_tools.yaml" />
+  <arg name="teleop_config" default="$(find moveit_jog_arm)/config/spacenav_via_teleop_tools.yaml" />
   <rosparam file="$(arg teleop_config)" command="load" />
 
   <node pkg="joy_teleop" type="joy_teleop.py" name="joy_teleop">
