+++ moveit_experimental/moveit_jog_arm/test/launch/jog_arm_integration_test.test
@@ -10,7 +10,7 @@
        These aren't suitable for actually jogging a robot in RViz, but they are good enough to 
        test jog node output.
   -->
-  <include file="$(find jog_arm)/test/arm_setup/launch/panda_move_group.launch">
+  <include file="$(find moveit_jog_arm)/test/arm_setup/launch/panda_move_group.launch">
     <arg name="fake_execution" value="true"/>
   </include>
 
@@ -19,20 +19,20 @@
   </node>
 
   <node name="joint_state_publisher" pkg="joint_state_publisher" type="joint_state_publisher">
-	  <rosparam command="load" file="$(find jog_arm)/test/arm_setup/config/initial_position.yaml" />
+	  <rosparam command="load" file="$(find moveit_jog_arm)/test/arm_setup/config/initial_position.yaml" />
   </node>
 
   <group ns="jog_arm_server" if="$(arg debug)">
-    <rosparam command="load" file="$(find jog_arm)/test/arm_setup/config/jog_settings.yaml"/>
+    <rosparam command="load" file="$(find moveit_jog_arm)/test/arm_setup/config/jog_settings.yaml"/>
   </group>
 
   <group unless="$(arg debug)">
-    <node name="jog_arm_server" pkg="jog_arm" type="jog_arm_server" output="screen">
+    <node name="jog_arm_server" pkg="moveit_jog_arm" type="jog_arm_server" output="screen">
       <param name="parameter_ns" type="string" value="jog_arm_server"/>
-      <rosparam command="load" file="$(find jog_arm)/test/arm_setup/config/jog_settings.yaml"/>
+      <rosparam command="load" file="$(find moveit_jog_arm)/test/arm_setup/config/jog_settings.yaml"/>
     </node>
   </group>
 
-  <param name="test_module" value="$(find jog_arm)/test/integration" />
+  <param name="test_module" value="$(find moveit_jog_arm)/test/integration" />
   <test test-name="jog_arm_integration_test" pkg="ros_pytest" type="ros_pytest_runner" />
 </launch>
